refactoring
remove save to local storage function
change ErrorMessage folder
change name of saved locations
