+++ src/Components/FeedbackPage/FeedbackForm/ErrorMessage/ErrorMessage.module.scss
@@ -1,5 +1,5 @@
-@import '../../../Styles/variables.scss';
-@import '../../../Styles/mixins.scss';
+@import '../../../../Styles/variables.scss';
+@import '../../../../Styles/mixins.scss';
 
 .dark-theme,
 .light-theme {
